Use TypicalBlockModel in SugTestUtil
